Fix pipefunctions pagination component
Move logic outside the dumb PaginatedList component
